% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllPrideProject.R
\name{getAllPrideProject}
\alias{getAllPrideProject}
\title{Minimal information available via PRIDE-api}
\usage{
getAllPrideProject(resultsPerPage = 100)
}
\description{
Get minimal information about PRIDE database
}
\examples{
prideProjects <- getAllPrideProject()
}
\author{
RINTU KUTUM
}
