% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrideProject.R
\name{getPrideProject}
\alias{getPrideProject}
\title{Get a Project information from PRIDE}
\usage{
getPrideProject(pxd)
}
\description{
Get a project information based on pride ID from PRIDE database
}
\examples{
pxd = "PXD004083"
pp <- getPrideProject(pxd)
}
\author{
RINTU KUTUM
}
