% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectCount.R
\name{getProjectCount}
\alias{getProjectCount}
\title{Get number of projects in PRIDE database}
\usage{
getProjectCount(query = NULL, speciesFilter = NULL, ptmsFilter = NULL,
  tissueFilter = NULL, diseaseFilter = NULL, titleFilter = NULL,
  instrumentFilter = NULL, experimentTypeFilter = NULL,
  quantificationFilter = NULL, projectTagFilter = NULL)
}
\description{
Get number of projects available in PRIDE
}
\examples{
nPrideProjects <- getProjectCount()
}
