% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectList.R
\name{getProjectList}
\alias{getProjectList}
\title{Get all Projects}
\usage{
getProjectList(term = NULL, resultsPerPage = 100, page = 0, sort = NULL,
  order = TRUE, taxonID = NULL, ptmsFilter = NULL, tissueFilter = NULL,
  diseaseFilter = NULL, titleFilter = NULL, instrumentFilter = NULL,
  experimentTypeFilter = NULL, quantificationFilter = NULL,
  projectTagFilter = NULL)
}
\description{
Get all projects lised in PRIDE database
}
