% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pleural2single}
\alias{pleural2single}
\title{Beautification codes for content2df PRIDE json results to data.frame}
\usage{
pleural2single(pleural)
}
\description{
Beautification codes for content2df PRIDE json results to data.frame
}
