% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prideMiniList2DF}
\alias{prideMiniList2DF}
\title{List to data.frame}
\usage{
prideMiniList2DF(prideMiniList)
}
\description{
getAllPrideProject helper function, converts list to data.frame format
}
\examples{
pxd = "PXD004083"
pp <- getPrideProject(pxd)
pp.df <- prideMiniList2DF(pp)
}
